mm: rename _count, field of the struct page, to _refcount
Many developers already know that field for reference count of the struct
page is _count and atomic type.  They would try to handle it directly and
this could break the purpose of page reference count tracepoint.  To
prevent direct _count modification, this patch rename it to _refcount and
add warning message on the code.  After that, developer who need to handle
reference count will find that field should not be accessed directly.

Signed-off-by: Joonsoo Kim <iamjoonsoo.kim@lge.com>
Cc: Hugh Dickins <hughd@google.com>
Cc: Johannes Berg <johannes@sipsolutions.net>
Cc: "David S. Miller" <davem@davemloft.net>
Cc: Sunil Goutham <sgoutham@cavium.com>
Cc: Chris Metcalf <cmetcalf@mellanox.com>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>