mm/slab: don't keep free slabs if free_objects exceeds free_limit
Currently, determination to free a slab is done whenever free object is
put into the slab.  This has a problem that free slabs are not freed even
if we have free slabs and have more free_objects than free_limit when
processed slab isn't a free slab.  This would cause to keep too much
memory in the slab subsystem.  This patch try to fix it by checking number
of free object after all free work is done.  If there is free slab at that
time, we can free it so we keep free slab as minimal as possible.

Signed-off-by: Joonsoo Kim <iamjoonsoo.kim@lge.com>
Cc: Christoph Lameter <cl@linux.com>
Cc: Pekka Enberg <penberg@kernel.org>
Cc: David Rientjes <rientjes@google.com>
Cc: Jesper Dangaard Brouer <brouer@redhat.com>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>