zsmalloc: squeeze inuse into page->mapping
Currently, we store class:fullness into page->mapping.  The number of
class we can support is 255 and fullness is 4 so (8 + 2 = 10bit) is enough
to represent them.  Meanwhile, the bits we need to store in-use objects in
zspage is that 11bit is enough.

For example, If we assume that 64K PAGE_SIZE, class_size 32 which is worst
case, class->pages_per_zspage become 1 so the number of objects in zspage
is 2048 so 11bit is enough.  The next class is 32 + 256(i.e.,
ZS_SIZE_CLASS_DELTA).  With worst case that ZS_MAX_PAGES_PER_ZSPAGE, 64K *
4 / (32 + 256) = 910 so 11bit is still enough.

So, we could squeeze inuse object count to page->mapping.

Signed-off-by: Minchan Kim <minchan@kernel.org>
Cc: Naoya Horiguchi <n-horiguchi@ah.jp.nec.com>
Cc: Vlastimil Babka <vbabka@suse.cz>
Cc: Joonsoo Kim <iamjoonsoo.kim@lge.com>
Cc: Konstantin Khlebnikov <koct9i@gmail.com>
Cc: Rafael Aquini <aquini@redhat.com>
Cc: Russ Knize <rknize@motorola.com>
Cc: Mel Gorman <mgorman@suse.de>
Cc: Hugh Dickins <hughd@google.com>
Cc: Sergey Senozhatsky <sergey.senozhatsky@gmail.com>
Cc: Rik van Riel <riel@redhat.com>
Cc: Gioh Kim <gi-oh.kim@profitbricks.com>
Cc: Sangseok Lee <sangseok.lee@lge.com>
Cc: Chan Gyun Jeong <chan.jeong@lge.com>
Cc: Al Viro <viro@ZenIV.linux.org.uk>
Cc: YiPing Xu <xuyiping@hisilicon.com>
Cc: Minchan Kim <minchan@kernel.org>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>