mm/compaction: support non-lru movable page migration
We have allowed migration for only LRU pages until now and it was enough
to make high-order pages.  But recently, embedded system(e.g., webOS,
android) uses lots of non-movable pages(e.g., zram, GPU memory) so we have
seen several reports about troubles of small high-order allocation.  For
fixing the problem, there were several efforts (e,g,.  enhance compaction
algorithm, SLUB fallback to 0-order page, reserved memory, vmalloc and so
on) but if there are lots of non-movable pages in system, their solutions
are void in the long run.

So, this patch is to support facility to change non-movable pages with
movable.  For the feature, this patch introduces functions related to
migration to address_space_operations as well as some page flags.

Basically, this patch supports two page-flags and two functions related to
page migration.  The flag and page->mapping stability are protected by
PG_lock.

	PG_movable
	PG_isolated

	bool (*isolate_page) (struct page *, isolate_mode_t);
	void (*putback_page) (struct page *);

Duty of subsystem want to make their pages as migratable are
as follows:

1. It should register address_space to page->mapping then mark the
   page as PG_movable via __SetPageMovable.

2. It should mark the page as PG_isolated via SetPageIsolated if
   isolation is sucessful and return true.

3. If migration is successful, it should clear PG_isolated and
   PG_movable of the page for free preparation then release the reference
   of the page to free.

4. If migration fails, putback function of subsystem should clear
   PG_isolated via ClearPageIsolated.

5. If a subsystem want to release isolated page, it should clear
   PG_isolated but not PG_movable.  Instead, VM will do it.

Signed-off-by: Gioh Kim <gurugio@hanmail.net>
Signed-off-by: Minchan Kim <minchan@kernel.org>
Cc: Vlastimil Babka <vbabka@suse.cz>
Cc: Mel Gorman <mgorman@suse.de>
Cc: Hugh Dickins <hughd@google.com>
Cc: Naoya Horiguchi <n-horiguchi@ah.jp.nec.com>
Cc: Joonsoo Kim <iamjoonsoo.kim@lge.com>
Cc: Konstantin Khlebnikov <koct9i@gmail.com>
Cc: Rafael Aquini <aquini@redhat.com>
Cc: Russ Knize <rknize@motorola.com>
Cc: Sergey Senozhatsky <sergey.senozhatsky@gmail.com>
Cc: Rik van Riel <riel@redhat.com>
Cc: Sangseok Lee <sangseok.lee@lge.com>
Cc: Chan Gyun Jeong <chan.jeong@lge.com>
Cc: Al Viro <viro@ZenIV.linux.org.uk>
Cc: YiPing Xu <xuyiping@hisilicon.com>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>