huge tmpfs: shrinker to migrate and free underused holes
Using 2MB for each small file is wasteful, and on average even a large
file is likely to waste 1MB at the end.  We could say that a huge tmpfs is
only suitable for huge files, but I would much prefer not to limit it in
that way, and would not be very able to test such a filesystem.

In our model, the unused space in the team is not put on any LRU (nor
charged to any memcg), so not yet accessible to page reclaim: we need a
shrinker to disband the team, and free up the unused space, under memory
pressure.  (Typically the freeable space is at the end, but there's no
assumption that it's at end of huge page or end of file.)

shmem_shrink_hugehole() is usually called from vmscan's shrink_slab(); but
I've found a direct call from shmem_alloc_page(), when it fails to
allocate a huge page (perhaps because too much memory is occupied by shmem
huge holes), is also helpful before a retry.

But each team holds a valuable resource: an extent of contiguous memory
that could be used for another team (or for an anonymous THP).  So try to
proceed in such a way as to conserve that resource: rather than just
freeing the unused space and leaving yet another huge page fragmented,
also try to migrate the used space to another partially occupied huge
page.

The algorithm in shmem_choose_hugehole() (find least occupied huge page in
older half of shrinklist, and migrate its cachepages into the most
occupied huge page with enough space to fit, again chosen from older half
of shrinklist) is unlikely to be ideal; but easy to implement as a
demonstration of the pieces which can be used by any algorithm, and good
enough for now.  A radix_tree tag helps to locate the partially occupied
huge pages more quickly: the tag available since shmem does not
participate in dirty/writeback accounting.

The "team_usage" field added to struct page (in union with "private") is
somewhat vaguely named: because while the huge page is sparsely occupied,
it counts the occupancy; but once the huge page is fully occupied, it will
come to be used differently in a later patch, as the huge mapcount (offset
by the HPAGE_PMD_NR occupancy) - it is never possible to map a sparsely
occupied huge page, because that would expose stale data to the user.

With this patch, the ShmemHugePages and ShmemFreeHoles lines of
/proc/meminfo are shown correctly; but ShmemPmdMapped remains 0.

Signed-off-by: Hugh Dickins <hughd@google.com>
Cc: "Kirill A. Shutemov" <kirill.shutemov@linux.intel.com>
Cc: Andrea Arcangeli <aarcange@redhat.com>
Cc: Andres Lagar-Cavilla <andreslc@google.com>
Cc: Yang Shi <yang.shi@linaro.org>
Cc: Ning Qu <quning@gmail.com>
Cc: David Rientjes <rientjes@google.com>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>